+++ init.lua
@@ -9,7 +9,7 @@
 
 local obj = {
     name = 'Zoom Join',
-    version = '3',
+    version = '4',
     author = 'Etan Reisner <deryni@gmail.com>',
     license = 'MPL-2.0',
     homepage = 'https://github.com/deryni/zoomjoin',
@@ -46,6 +46,7 @@ obj.editmenu = true
 
 -- The zoom application's bundleID
 local bundleID = 'us.zoom.xos'
+local appname = 'zoom.us'
 
 -- The zoom menu
 local zoommenu = nil
@@ -309,7 +310,7 @@ function obj.joinMeeting(urlorid, password)
         return
     end
 
-    hs.application.find('zoom.us'):activate()
+    hs.application.launchOrFocus(appname)
 
     if urlorid:match('^https://') then
         return hs.urlevent.openURLWithBundle(urlorid, bundleID)