+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Version 4 (####-##-##)
+
+### Features
+* Launch zoom if it isn't running
+
 ## Version 3 (2022-05-27)
 
 ### Features